Library: Consolidate has_footprint implementations
