+++ src/faebryk/library/_F.py
@@ -82,15 +82,16 @@ from faebryk.library.Relay import Relay
 from faebryk.library.LogicOps import LogicOps
 from faebryk.library.can_attach_to_footprint import can_attach_to_footprint
 from faebryk.library.can_attach_via_pinmap import can_attach_via_pinmap
-from faebryk.library.has_footprint_impl import has_footprint_impl
+from faebryk.library.has_footprint_defined import has_footprint_defined
 from faebryk.library.has_kicad_footprint import has_kicad_footprint
 from faebryk.library.Pad import Pad
 from faebryk.library.Button import Button
 from faebryk.library.GDT import GDT
 from faebryk.library.has_pin_association_heuristic_lookup_table import has_pin_association_heuristic_lookup_table
 from faebryk.library.LogicGate import LogicGate
-from faebryk.library.has_footprint_defined import has_footprint_defined
+from faebryk.library.can_attach_to_footprint_via_pinmap import can_attach_to_footprint_via_pinmap
 from faebryk.library.Net import Net
+from faebryk.library.can_attach_to_footprint_symmetrically import can_attach_to_footprint_symmetrically
 from faebryk.library.can_attach_via_pinmap_pinlist import can_attach_via_pinmap_pinlist
 from faebryk.library.has_equal_pins import has_equal_pins
 from faebryk.library.has_kicad_manual_footprint import has_kicad_manual_footprint
@@ -99,32 +100,22 @@ from faebryk.library.BJT import BJT
 from faebryk.library.Diode import Diode
 from faebryk.library.MOSFET import MOSFET
 from faebryk.library.LogicGates import LogicGates
-from faebryk.library.can_attach_to_footprint_symmetrically import can_attach_to_footprint_symmetrically
-from faebryk.library.can_attach_to_footprint_via_pinmap import can_attach_to_footprint_via_pinmap
+from faebryk.library.B4B_ZR_SM4_TF import B4B_ZR_SM4_TF
+from faebryk.library.DE9Connector import DE9Connector
+from faebryk.library.pf_533984002 import pf_533984002
 from faebryk.library.has_pcb_routing_strategy_manual import has_pcb_routing_strategy_manual
 from faebryk.library.has_pcb_routing_strategy_via_to_layer import has_pcb_routing_strategy_via_to_layer
-from faebryk.library.can_attach_via_pinmap_equal import can_attach_via_pinmap_equal
-from faebryk.library.has_equal_pins_in_ifs import has_equal_pins_in_ifs
-from faebryk.library.has_kicad_footprint_equal_ifs import has_kicad_footprint_equal_ifs
-from faebryk.library.KicadFootprint import KicadFootprint
-from faebryk.library.TVS import TVS
 from faebryk.library.Capacitor import Capacitor
 from faebryk.library.Crystal import Crystal
 from faebryk.library.Fuse import Fuse
 from faebryk.library.Inductor import Inductor
 from faebryk.library.Resistor import Resistor
 from faebryk.library.Switch import Switch
-from faebryk.library.B4B_ZR_SM4_TF import B4B_ZR_SM4_TF
-from faebryk.library.DE9Connector import DE9Connector
-from faebryk.library.pf_533984002 import pf_533984002
-from faebryk.library.DIP import DIP
-from faebryk.library.QFN import QFN
-from faebryk.library.SMDTwoPin import SMDTwoPin
-from faebryk.library.SOIC import SOIC
-from faebryk.library.has_kicad_footprint_equal_ifs_defined import has_kicad_footprint_equal_ifs_defined
-from faebryk.library.Mounting_Hole import Mounting_Hole
-from faebryk.library.can_be_surge_protected import can_be_surge_protected
-from faebryk.library.is_surge_protected import is_surge_protected
+from faebryk.library.can_attach_via_pinmap_equal import can_attach_via_pinmap_equal
+from faebryk.library.has_equal_pins_in_ifs import has_equal_pins_in_ifs
+from faebryk.library.has_kicad_footprint_equal_ifs import has_kicad_footprint_equal_ifs
+from faebryk.library.KicadFootprint import KicadFootprint
+from faebryk.library.TVS import TVS
 from faebryk.library.MultiCapacitor import MultiCapacitor
 from faebryk.library.can_be_decoupled import can_be_decoupled
 from faebryk.library.is_decoupled import is_decoupled
@@ -133,13 +124,21 @@ from faebryk.library.DifferentialPair import DifferentialPair
 from faebryk.library.Potentiometer import Potentiometer
 from faebryk.library.ResistorVoltageDivider import ResistorVoltageDivider
 from faebryk.library.Resistor_Voltage_Divider import Resistor_Voltage_Divider
-from faebryk.library.is_surge_protected_defined import is_surge_protected_defined
+from faebryk.library.DIP import DIP
+from faebryk.library.QFN import QFN
+from faebryk.library.SMDTwoPin import SMDTwoPin
+from faebryk.library.SOIC import SOIC
+from faebryk.library.has_kicad_footprint_equal_ifs_defined import has_kicad_footprint_equal_ifs_defined
+from faebryk.library.Mounting_Hole import Mounting_Hole
+from faebryk.library.can_be_surge_protected import can_be_surge_protected
+from faebryk.library.is_surge_protected import is_surge_protected
 from faebryk.library.is_decoupled_nodes import is_decoupled_nodes
 from faebryk.library.Ethernet import Ethernet
 from faebryk.library.RS485 import RS485
 from faebryk.library.USB_Type_C_Receptacle_24_pin import USB_Type_C_Receptacle_24_pin
-from faebryk.library.can_be_surge_protected_defined import can_be_surge_protected_defined
+from faebryk.library.is_surge_protected_defined import is_surge_protected_defined
 from faebryk.library.can_be_decoupled_defined import can_be_decoupled_defined
+from faebryk.library.can_be_surge_protected_defined import can_be_surge_protected_defined
 from faebryk.library.ElectricPower import ElectricPower
 from faebryk.library.B0505S_1WR3 import B0505S_1WR3
 from faebryk.library.Battery import Battery
